requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.1
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
click==7.1.2
discord.py==1.3.3
gTTS==2.1.1
gTTS-token==1.1.3
idna==2.9
multidict==4.7.6
pycparser==2.20
PyNaCl==1.4.0
python-dotenv==0.13.0
redis==3.5.3
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.9
websockets==8.1
yarl==1.4.2
